Bug fixes and new method in Board.class
Added getPartialQuarto(Piece[][] board, char type) to Board. This allows you to get an ArrayList of Strings that represent partial Quartos. You can either look for halfQuartos or almostQuartos. This will allow you to test how placing a piece will change the board without having to edit the original Board.

checkForQuarto contained a bug where it was putting rows that weren't valid half Quartos into the halfQuarto ArrayLists.
This was because of how Piece.comparePieces(Piece[] p) and Board.checkForQuartos() works. The String returned by comparePieces() is five characters long, and the first character contains the number of Pieces in the comparison. When comparing two Pieces, the first character is always '2' (to say taht two pieces exist in the line of Pieces being examined). However, this conflicted with the fact that when checkForQuartos() sees the character '2' it takes that as a sign that the line being examined contains a partial or full Quarto.
So if you had a line of two Pieces that share no traits, comparePieces would return this: "20000"
but when checkForQuarto sees that, it'll count it as a halfQuarto because it only sees the '2' even though the 2 isn't in the correct spot for it to count as a partial Quarto.

That was longwinded but tl:dr; I fixed it and it should work correctly now.