Edits to early game heuristics
getPiece() edited to better return a piece and to prevent issue with AI attempting to double return a piece. I left in the code or the case where it warns you if a piece was not returned because I found this useful for debugging and ideally it should never be executed anyway.

numPiecesOnBoardWithAttribute(int attribute) straight up wasn't working due to misuse of values and logic errors while comparing ints with characters, causing the method to fail and causing a random piece to be returned during early game.

getOptimalSpace(Piece[][] board, int numPiecesPerLine) wasn't functioning correctly, fixed it.

choosePiece(Piece[] remainingPieces) wasn't working due to both internal logic errors and numPiecesOnBoardWithAttribute(int attribute) not working. It functions properly now