Bug patch
Occasionally the game would infinitely loop during brute force because the AI kept attempting to give a piece that was already on the board. This small change is to try and prevent that by forcing var pieceToGive to be null after the piece in it has been given so that it always goes through noQuartoGetPiece() to get its piece.